Update
Improved the code and added new things.
- Added class: LinksPanel
- Added class: VsPanel

Made the building blocks for the reddit / HLTV feature, with which you
should be able to look at reddit or hltv for information on the matches
you want to bet on.

Will add more comments to the code, as it might seem confusing now.
Probably some things messed up, but we'll fix that.

Explanation of 2 new classes:
VsPanel contains the label you're going to click on to get to the page
you want information from, it will change it's mouseListener to send you
to different pages.
LinksPanel just contains the links to the mainpages (csgolounge.com,
htlv.org, reddit.com/r/csgobetting) and displays them.